Complete the GF Kernel using glyphs from EC fonts
This commit is big because I switched to use the EC fonts [1] for
pulling the missing characters instead of relying on the Computer Modern
Math Symbols and Computer Modern Math Italic.

Here are some reasons for this decision:

- like Computer Modern, the EC fonts also use METAFONT so processing
  pipeline stays the same.
- I haven't check each glyph but the METAFONT code seems to be identical
  for most of the ovelapping glpyhs.
- the EC fonts have an atypical but permissive license.

EC comes with an old and unusual euro currency sign so this commit also
pulls that instead from the euro-ce [3] METAFONT font that is covered by
the BSD license.

Lastly, the U+2026 HORIZONTAL ELLIPSIS was created by using FontForge
Python API calls that manipulate the "period" glyph.

And one last-last thing: this commit also replaces the sources/build.sh
with a Makefile and extends the root Makefile to make it easier to find
the missing glyphs.

This commit should allow closing issue #1.

Current state: 158 glyphs for both Regular and Italic.

[1] https://ctan.org/tex-archive/fonts/ec
[2] http://mirrors.ctan.org/fonts/ec/src/copyrite.txt
[3] https://ctan.org/tex-archive/fonts/euro-ce