Switch to use FontForge for loading and tracing the .gf files (#9)
This commit replaces mftrace with the functionality existing in
FontForge. This commit also switches to only using the EC and TC fonts
instead of the Computer Modern because it makes things simpler. All the
TC glpyhs that are not in EC are also pulled in. As before, the Euro
symbol comes from the euro-ce CTAN package.

This PR also switches to 10000 DPI bitmaps.

There is some more cleanup work to do like removing the mftrace and
updating the specimens. It's late and I need to go home so I'll do that
in a separate PR.