Add the last 3 remaining glyphs to complete GF Latin Core (#21)
This PR adds 3 more glyphs:

 - U+0110 LATIN CAPITAL LETTER H WITH STROKE
 - U+0126 LATIN CAPITAL LETTER H WITH STROKE
 - U+0127 LATIN SMALL LETTER H WITH STROKE

This PR is adding these glyphs via METAFONT because a similar glyph,
U+0111 (LATIN SMALL LETTER D WITH STROKE)) is already present in EC as
"The letter bar d" on position 236 in octal and 96 in hexa. The added
glyphs probably need more tuning and I noted in a TODO one issue I
noticed.

This completes adding the support for GF Latin Core from #12. 🎉